+++ Lab10/Lab10/db4odemo/src/userinterface/PatientRole/RequestMedicinesJPanel.java
@@ -301,7 +301,7 @@ public class RequestMedicinesJPanel extends javax.swing.JPanel {
 
     private void AddtoCartActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_AddtoCartActionPerformed
         // TODO add your handling code here:
-         int Price = 0, updatedPrice = 0;
+         float Price = 0, updatedPrice = 0;
        //  ArrayList<String> list = new ArrayList<>();
 
         Medicine m = (Medicine) ChooseMedCombo.getSelectedItem();
@@ -309,7 +309,7 @@ public class RequestMedicinesJPanel extends javax.swing.JPanel {
         int qty = Integer.parseInt(Qtyxt.getText());
         for (Medicine medi : med.getMedicineList()) {
             if (m.getSaltname().equals(medi.getSaltname())) {
-                Price = medi.getPrice() * qty;
+                Price = (float) (medi.getPrice() * qty);
                 Pricetxt.setText(String.valueOf(Price) + "$");
                 Pricetxt.setEnabled(false);
             }
@@ -332,7 +332,7 @@ public class RequestMedicinesJPanel extends javax.swing.JPanel {
 
     private void BuyNowActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_BuyNowActionPerformed
         // TODO add your handling code here:
-        int Price=0;
+        float Price=0;
         finallist =new ArrayList<>();
         try{
         Medicine m = (Medicine) ChooseMedCombo.getSelectedItem();
@@ -370,7 +370,7 @@ public class RequestMedicinesJPanel extends javax.swing.JPanel {
         // TODO add your handling code here:
         
         Medicine medicine= (Medicine) ChooseMedCombo.getSelectedItem();
-        int Price=0;
+        float Price=0;
          for(Medicine medi : med.getMedicineList())
          {
              if(medicine.getSaltname().equals(medi.getSaltname())){
